Create a map with the risk color of each region (NUTS3 granularity), without financial data.

Create a map with the risk color of each region (NUTS3 granularity), without financial data.

Usage

map_region_risk(
  data,
  country_code = c("DE"),
  grouping_emission = grouping_emission(),
  mode = modes(),
  scenario = scenarios(),
  year = years(),
  risk_category = risk_category()
)

Arguments

data

A data frame like without_financial

country_code

Country code (ISO 3166 alpha-2) for which the map will be plotted.

grouping_emission

The mode of benchmark to plot. It can be one of "all", "unit" or "tilt_sector", "unit_tilt_sector", "isic_4digit" or "unit_isic_4digit". If nothing is chosen, "all" is the default mode.

mode

The mode to plot. It can be one of "equal_weight", "worst_case" or "best_case". If nothing is chosen, "equal_weight" is the default mode.

scenario

A character vector: 1.5C RPS, NZ 2050.

year

A character vector: 2030, 2050.

risk_category

A character vector.

Value

A ggplot2 object representing the country data plot.

Examples

# Plot a German with a "unit" benchmark and equal_weight finance
try({
  map_region_risk(without_financial, country_code = "DE", benchmark = "unit")
})
#> Error in map_region_risk(without_financial, country_code = "DE", benchmark = "unit") : 
#>   unused argument (benchmark = "unit")